+++ src/Umbraco.Web/Mvc/UmbracoTemplatePage.cs
@@ -71,46 +71,5 @@ namespace Umbraco.Web.Mvc
                                                  : new UmbracoHelper(UmbracoContext, Model.Content));
             }
         }
-
-        /// <summary>
-        /// This will detect the end /body tag and insert the preview badge if in preview mode
-        /// </summary>
-        /// <param name="value"></param>
-        public override void WriteLiteral(object value)
-        {
-            // filter / add preview banner
-            if (Response.ContentType.InvariantEquals("text/html")) // ASP.NET default value
-            {
-
-                if (UmbracoContext.Current.IsDebug || UmbracoContext.Current.InPreviewMode)
-                {
-                    var text = value.ToString().ToLowerInvariant();
-                    int pos = text.IndexOf("</body>");
-                    if (pos > -1)
-                    {
-                        if (UmbracoContext.Current.InPreviewMode)
-                        {
-                            var htmlBadge =
-                                String.Format(UmbracoSettings.PreviewBadge,
-                                    IOHelper.ResolveUrl(SystemDirectories.Umbraco),
-                                    IOHelper.ResolveUrl(SystemDirectories.UmbracoClient),
-                                    Server.UrlEncode(UmbracoContext.Current.HttpContext.Request.Path));
-
-                            text = text.Substring(0, pos) + htmlBadge + text.Substring(pos, text.Length - pos);
-                        }
-                        else
-                        {
-                            var profilerMarkup = this.Html.RenderProfiler();
-                            text = text.Substring(0, pos) + profilerMarkup + text.Substring(pos, text.Length - pos);
-                        }
-                        base.WriteLiteral(text);
-                        return;
-                    }
-                }
-            }
-
-            base.WriteLiteral(value);
-        }
-
     }
 }
